+++ fastapi_mvc_template/app/asgi.py
@@ -3,15 +3,12 @@
 import logging
 
 from fastapi import FastAPI
-from fastapi_mvc_template.app.config.router import router
-from fastapi_mvc_template.app.config.application import (
-    DEBUG,
-    PROJECT_NAME,
-    VERSION,
+from fastapi_mvc_template.config import router, settings
+from fastapi_mvc_template.app.utils import RedisClient, AiohttpClient
+from fastapi_mvc_template.app.exceptions import (
+    HTTPException,
+    http_exception_handler,
 )
-from fastapi_mvc_template.app.utils.redis import RedisClient
-from fastapi_mvc_template.app.utils.aiohttp_client import AiohttpClient
-
 
 log = logging.getLogger(__name__)
 
@@ -19,25 +16,29 @@ log = logging.getLogger(__name__)
 async def on_startup():
     """Fastapi startup event handler.
 
-    Creates AiohttpClient session.
+    Creates RedisClient and AiohttpClient session.
 
     """
     log.debug("Execute FastAPI startup event handler.")
     # Initialize utilities for whole FastAPI application without passing object
     # instances within the logic. Feel free to disable it if you don't need it.
-    RedisClient.open_redis_client()
+    if settings.USE_REDIS:
+        await RedisClient.open_redis_client()
+
     AiohttpClient.get_aiohttp_client()
 
 
 async def on_shutdown():
     """Fastapi shutdown event handler.
 
-    Destroys AiohttpClient session.
+    Destroys RedisClient and AiohttpClient session.
 
     """
     log.debug("Execute FastAPI shutdown event handler.")
     # Gracefully close utilities.
-    await RedisClient.close_redis_client()
+    if settings.USE_REDIS:
+        await RedisClient.close_redis_client()
+
     await AiohttpClient.close_aiohttp_client()
 
 
@@ -50,15 +51,17 @@ def get_app():
     """
     log.debug("Initialize FastAPI application node.")
     app = FastAPI(
-        title=PROJECT_NAME,
-        debug=DEBUG,
-        version=VERSION,
-        docs_url="/",
+        title=settings.PROJECT_NAME,
+        debug=settings.DEBUG,
+        version=settings.VERSION,
+        docs_url=settings.DOCS_URL,
         on_startup=[on_startup],
         on_shutdown=[on_shutdown],
     )
     log.debug("Add application routes.")
     app.include_router(router)
+    # Register global exception handler for custom HTTPException.
+    app.add_exception_handler(HTTPException, http_exception_handler)
 
     return app