+++ fastapi_mvc_template/app/exceptions/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""FastAPI MVC template."""
+from .http import HTTPException, http_exception_handler
+
+
+__all__ = (HTTPException, http_exception_handler)