+++ fastapi_mvc_template/app/models/ready.py
@@ -1,10 +1,12 @@
 # -*- coding: utf-8 -*-
 """Ready model."""
+from typing import Any, Dict
+
 from pydantic import BaseModel
 
 
 class ReadyResponse(BaseModel):
-    """Ready response model.
+    """Ready response model definition.
 
     Attributes:
         status(str): Strings are accepted as-is, int float and Decimal are
@@ -19,3 +21,27 @@ class ReadyResponse(BaseModel):
     """
 
     status: str
+
+    class Config:
+        """Config sub-class needed to extend/override the generated JSON schema.
+
+        More details can be found in pydantic documentation:
+        https://pydantic-docs.helpmanual.io/usage/schema/#schema-customization
+
+        """
+
+        @staticmethod
+        def schema_extra(schema: Dict[str, Any]) -> None:
+            """Post-process the generated schema.
+
+            Mathod can have one or two positional arguments. The first will be
+            the schema dictionary. The second, if accepted, will be the model
+            class. The callable is expected to mutate the schema dictionary
+            in-place; the return value is not used.
+
+            Args:
+                schema(Dict[str, Any]): The schema dictionary.
+
+            """
+            # Override schema description, by default is taken from docstring.
+            schema["description"] = "Ready response model."