+++ fastapi_mvc_template/app/utils/__init__.py
@@ -1,2 +1,9 @@
 # -*- coding: utf-8 -*-
 """FastAPI MVC template."""
+from .aiohttp_client import AiohttpClient
+from .redis import RedisClient
+
+__all__ = (
+    AiohttpClient,
+    RedisClient,
+)