+++ fastapi_mvc_template/app/utils/redis.py
@@ -5,13 +5,7 @@ import logging
 import aioredis
 import aioredis.sentinel
 from aioredis.exceptions import RedisError
-from fastapi_mvc_template.app.config.redis import (
-    REDIS_HOST,
-    REDIS_PORT,
-    REDIS_USERNAME,
-    REDIS_PASSWORD,
-    REDIS_USE_SENTINEL
-)
+from fastapi_mvc_template.config import redis as redis_conf
 
 
 class RedisClient(object):
@@ -32,7 +26,7 @@ class RedisClient(object):
     log: logging.Logger = logging.getLogger(__name__)
     base_redis_init_kwargs: dict = {
         "encoding": "utf-8",
-        "port": REDIS_PORT
+        "port": redis_conf.REDIS_PORT,
     }
     connection_kwargs: dict = {}
 
@@ -48,22 +42,22 @@ class RedisClient(object):
         """
         if cls.redis_client is None:
             cls.log.debug("Initialize Redis client.")
-            if REDIS_USERNAME and REDIS_PASSWORD:
+            if redis_conf.REDIS_USERNAME and redis_conf.REDIS_PASSWORD:
                 cls.connection_kwargs = {
-                    "username": REDIS_USERNAME,
-                    "password": REDIS_PASSWORD,
+                    "username": redis_conf.REDIS_USERNAME,
+                    "password": redis_conf.REDIS_PASSWORD,
                 }
 
-            if REDIS_USE_SENTINEL:
+            if redis_conf.REDIS_USE_SENTINEL:
                 sentinel = aioredis.sentinel.Sentinel(
-                    [(REDIS_HOST, REDIS_PORT)],
-                    sentinel_kwargs=cls.connection_kwargs
+                    [(redis_conf.REDIS_HOST, redis_conf.REDIS_PORT)],
+                    sentinel_kwargs=cls.connection_kwargs,
                 )
                 cls.redis_client = sentinel.master_for("mymaster")
             else:
                 cls.base_redis_init_kwargs.update(cls.connection_kwargs)
                 cls.redis_client = aioredis.from_url(
-                    "redis://{0:s}".format(REDIS_HOST),
+                    "redis://{0:s}".format(redis_conf.REDIS_HOST),
                     **cls.base_redis_init_kwargs,
                 )
 
@@ -98,9 +92,9 @@ class RedisClient(object):
         except RedisError as ex:
             cls.log.exception(
                 "Redis PING command finished with exception",
-                exc_info=(type(ex), ex, ex.__traceback__)
+                exc_info=(type(ex), ex, ex.__traceback__),
             )
-            raise ex
+            return False
 
     @classmethod
     async def set(cls, key, value):
@@ -131,7 +125,7 @@ class RedisClient(object):
         except RedisError as ex:
             cls.log.exception(
                 "Redis SET command finished with exception",
-                exc_info=(type(ex), ex, ex.__traceback__)
+                exc_info=(type(ex), ex, ex.__traceback__),
             )
             raise ex
 
@@ -165,7 +159,7 @@ class RedisClient(object):
         except RedisError as ex:
             cls.log.exception(
                 "Redis RPUSH command finished with exception",
-                exc_info=(type(ex), ex, ex.__traceback__)
+                exc_info=(type(ex), ex, ex.__traceback__),
             )
             raise ex
 
@@ -195,7 +189,7 @@ class RedisClient(object):
         except RedisError as ex:
             cls.log.exception(
                 "Redis EXISTS command finished with exception",
-                exc_info=(type(ex), ex, ex.__traceback__)
+                exc_info=(type(ex), ex, ex.__traceback__),
             )
             raise ex
 
@@ -219,15 +213,13 @@ class RedisClient(object):
         """
         redis_client = cls.redis_client
 
-        cls.log.debug(
-            "Preform Redis GET command, key: {}".format(key)
-        )
+        cls.log.debug("Preform Redis GET command, key: {}".format(key))
         try:
             return await redis_client.get(key)
         except RedisError as ex:
             cls.log.exception(
                 "Redis GET command finished with exception",
-                exc_info=(type(ex), ex, ex.__traceback__)
+                exc_info=(type(ex), ex, ex.__traceback__),
             )
             raise ex
 
@@ -268,6 +260,6 @@ class RedisClient(object):
         except RedisError as ex:
             cls.log.exception(
                 "Redis LRANGE command finished with exception",
-                exc_info=(type(ex), ex, ex.__traceback__)
+                exc_info=(type(ex), ex, ex.__traceback__),
             )
             raise ex