+++ tests/unit/app/conftest.py
@@ -1,9 +1,14 @@
 import pytest
 from fastapi.testclient import TestClient
 from fastapi_mvc_template.app.asgi import get_app
+from fastapi_mvc_template.config import settings
 
 
 @pytest.fixture
 def app():
+    # Overriding to true in order to initialize redis client on FastAPI event
+    # startup handler. It'll be needed for unit tests.
+    settings.USE_REDIS = True
     app = get_app()
-    yield TestClient(app)
+    with TestClient(app) as client:
+        yield client